ADDED: Support for Chinese file names in msys2 environment.
The Perl package that comes with MSYS2 calls the Windows narrow character API (ANSI API).
This API does not recognize UTF8, but uses the system default code page.
(On Chinese Windows this is CP936/GBK).
Therefore, the Chinese text in the file name is passed in in the CP936
parameter format, and an error will occur when decoding it as UTF8.

In the Chinese environment of msys2, When comparing whether a directory exists,
we can only use the cp936 format for comparison, using utf8 will be ineffective.